+++ sentry-android-core/src/main/java/io/sentry/android/core/AppLifecycleIntegration.java
@@ -42,7 +42,7 @@ public final class AppLifecycleIntegration implements Integration, Closeable {
         .log(
             SentryLevel.DEBUG,
             "enableSessionTracking enabled: %s",
-            this.options.isEnableSessionTracking());
+            this.options.isEnableAutoSessionTracking());
 
     this.options
         .getLogger()
@@ -51,7 +51,8 @@ public final class AppLifecycleIntegration implements Integration, Closeable {
             "enableAppLifecycleBreadcrumbs enabled: %s",
             this.options.isEnableAppLifecycleBreadcrumbs());
 
-    if (this.options.isEnableSessionTracking() || this.options.isEnableAppLifecycleBreadcrumbs()) {
+    if (this.options.isEnableAutoSessionTracking()
+        || this.options.isEnableAppLifecycleBreadcrumbs()) {
       try {
         Class.forName("androidx.lifecycle.DefaultLifecycleObserver");
         Class.forName("androidx.lifecycle.ProcessLifecycleOwner");
@@ -82,16 +83,26 @@ public final class AppLifecycleIntegration implements Integration, Closeable {
         new LifecycleWatcher(
             hub,
             this.options.getSessionTrackingIntervalMillis(),
-            this.options.isEnableSessionTracking(),
+            this.options.isEnableAutoSessionTracking(),
             this.options.isEnableAppLifecycleBreadcrumbs());
     ProcessLifecycleOwner.get().getLifecycle().addObserver(watcher);
     options.getLogger().log(SentryLevel.DEBUG, "AppLifecycleIntegration installed.");
   }
 
+  private void removeObserver() {
+    ProcessLifecycleOwner.get().getLifecycle().removeObserver(watcher);
+  }
+
   @Override
   public void close() throws IOException {
     if (watcher != null) {
-      ProcessLifecycleOwner.get().getLifecycle().removeObserver(watcher);
+      if (MainThreadChecker.isMainThread()) {
+        removeObserver();
+      } else {
+        // some versions of the androidx lifecycle-process require this to be executed on the main
+        // thread.
+        handler.post(() -> removeObserver());
+      }
       watcher = null;
       if (options != null) {
         options.getLogger().log(SentryLevel.DEBUG, "AppLifecycleIntegration removed.");