+++ sentry/src/main/java/io/sentry/util/CollectionUtils.java
@@ -1,6 +1,7 @@
 package io.sentry.util;
 
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import org.jetbrains.annotations.ApiStatus;
@@ -44,5 +45,34 @@ public final class CollectionUtils {
     } else {
       return null;
     }
+  }
+
+  /**
+   * Returns a new map which entries match a predicate specified by a parameter.
+   *
+   * @param map - the map to filter
+   * @param predicate - the predicate
+   * @param <K> - map entry key type
+   * @param <V> - map entry value type
+   * @return a new map
+   */
+  public static @NotNull <K, V> Map<K, V> filterMapEntries(
+      final @NotNull Map<K, V> map, final @NotNull Predicate<Map.Entry<K, V>> predicate) {
+    final Map<K, V> filteredMap = new HashMap<>();
+    for (final Map.Entry<K, V> entry : map.entrySet()) {
+      if (predicate.test(entry)) {
+        filteredMap.put(entry.getKey(), entry.getValue());
+      }
+    }
+    return filteredMap;
+  }
+
+  /**
+   * A simplified copy of Java 8 Predicate.
+   *
+   * @param <T> the type
+   */
+  public interface Predicate<T> {
+    boolean test(T t);
   }
 }